backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.6
python-multipart>=0.0.9
Pillow>=10.4.0
PyMuPDF>=1.24.10
numpy>=2.1.1
easyocr>=1.7.0
google-cloud-vision>=3.7.0
